Test for identifier address
